Move isValidTest to module
